Continued refactor of model-specific data augmentation
